+++ pennylane_lightning/core/src/simulators/lightning_qubit/gates/cpu_kernels/GateImplementationsLM.hpp
@@ -154,6 +154,7 @@ class GateImplementationsLM : public PauliGenerator<GateImplementationsLM> {
         GateOperation::DoubleExcitationPlus,
         GateOperation::MultiRZ,
         GateOperation::GlobalPhase,
+        GateOperation::PCPhase,
     };
 
     constexpr static std::array implemented_controlled_gates = {
@@ -182,6 +183,7 @@ class GateImplementationsLM : public PauliGenerator<GateImplementationsLM> {
         ControlledGateOperation::DoubleExcitationPlus,
         ControlledGateOperation::MultiRZ,
         ControlledGateOperation::GlobalPhase,
+        ControlledGateOperation::PCPhase,
     };
 
     constexpr static std::array implemented_generators = {
@@ -1999,6 +2001,7 @@ class GateImplementationsLM : public PauliGenerator<GateImplementationsLM> {
         applyNCN(arr, num_qubits, controlled_wires, controlled_values, wires,
                  core_function);
     }
+
     template <class PrecisionT, class ParamT>
     static void
     applyGlobalPhase(std::complex<PrecisionT> *arr, std::size_t num_qubits,
@@ -2041,6 +2044,58 @@ class GateImplementationsLM : public PauliGenerator<GateImplementationsLM> {
             applyNC1<PrecisionT, PrecisionT, decltype(core_function), true>(
                 arr, num_qubits, controlled_wires, controlled_values, {target},
                 core_function);
+        }
+    }
+
+    template <class PrecisionT, class ParamT>
+    static void applyPCPhase(std::complex<PrecisionT> *arr,
+                             std::size_t num_qubits,
+                             const std::vector<std::size_t> &wires,
+                             bool inverse, ParamT angle, ParamT dim) {
+        applyNCPCPhase(arr, num_qubits, {}, {}, wires, inverse, angle, dim);
+    }
+
+    template <class PrecisionT, class ParamT>
+    static void applyNCPCPhase(std::complex<PrecisionT> *arr,
+                               std::size_t num_qubits,
+                               const std::vector<std::size_t> &controlled_wires,
+                               const std::vector<bool> &controlled_values,
+                               const std::vector<std::size_t> &wires,
+                               bool inverse, ParamT angle, ParamT dim) {
+
+        const auto dim_size_t = static_cast<std::size_t>(std::round(dim));
+
+        PL_ABORT_IF(dim_size_t < 0 || dim_size_t > 1U << num_qubits,
+                    "The dimension of the PCPhase gate must be a positive "
+                    "integer and less than or equal to statevector size.");
+
+        const PrecisionT phase = inverse ? -angle : angle;
+        const std::complex<PrecisionT> upper_complex = {std::cos(phase),
+                                                        std::sin(phase)};
+        const std::complex<PrecisionT> lower_complex = std::conj(upper_complex);
+
+        // Matrix
+        const std::size_t matrix_dim = Pennylane::Util::exp2(wires.size());
+
+        std::vector<std::complex<PrecisionT>> matrixPCPhase(matrix_dim *
+                                                            matrix_dim);
+
+        // Fill diagonal
+        for (std::size_t i = 0; i < dim_size_t; i++) {
+            matrixPCPhase[i * matrix_dim + i] = upper_complex;
+        }
+        for (std::size_t i = dim_size_t; i < matrix_dim; i++) {
+            matrixPCPhase[i * matrix_dim + i] = lower_complex;
+        }
+
+        if (controlled_wires.empty()) {
+            applyMultiQubitOp<PrecisionT>(arr, num_qubits, matrixPCPhase.data(),
+                                          wires, false);
+
+        } else {
+            applyNCMultiQubitOp<PrecisionT>(
+                arr, num_qubits, matrixPCPhase.data(), controlled_wires,
+                controlled_values, wires, false);
         }
     }