+++ .github/CHANGELOG.md
@@ -15,6 +15,9 @@
 
 ### Improvements
 
+* Add native support for `qml.PCPhase` for `lightning.qubit` and `lightning.gpu`.
+  [(#1107)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1107)
+
 * Use cuquantum API for calculating expectation value of Pauli sentences in Lightning-GPU.
   [(#1104)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1104)